Make sure that editor window starts at the first column.
Programs like git output a line to stderr before starting the
editor. Make sure that we clear this line so that we always start
at the first column. Later, we should find a sensible way for the
user to see the message from git as well as start our editor at the
first column.