Support suspend and resume using shell's job control.
ted does not support multiple buffers and probably never will.
However, it is nice to be able to edit multiple files concurrently.
You can use C-z to suspend ted and later use `fg' to resume it.

You can only suspend at the beginning of the edit loop. That is,
this is not allowed:
- Type C-x to activate extended keymap.
- Type C-z to suspend ted.
- Resume and type C-s to save the file.

This behaviour is consistent with other terminal screen editors
such as vim.