Refactor save_buffer().
A macro for_each_block(buf, sz, i, body) is added that fills buf
with at most sz next bytes from the editor buffer and executes body
with i set to the number of actual bytes in buf. This is done until
all bytes in the editor buffer is processed.