save_buffer(): make it safer.
We don't directly write to the file upon saving.
- Write contents to a temporary file.
- Rename temporary file to overwrite the file.

If something goes wrong during writing to the temporary file, the
old contents are preserved. If something goes wrong after writing
to the temporary file, we keep both and notify the user.

Also, before renaming, we check whether the original file has been
modified since the last write (or the first read). If so, we don't
rename and notify the user about the temporary file.