Keychords to manage indentation.
`M-j` and `M-J` increases and decreases the indentation of the
current line by one unit.

The indentation unit can be specified using the `-i` command-line
argument. Its default value is a <Tab>.

For implementing this, we also add stable marks and temporary
marks. A stable mark keeps its position in the text even after
insertion or deletion in the buffer. We do this by adding two
functions to update all marks after insertion and deletion.

Temporary marks are not visible to the user. They are only for
use by the editor. Their creation and destruction must follow
a stack discipline.